Add attribution for stack parsers
